destroyアクション変更

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, only: :new
+  before_action :authenticate_user!, only: [:new, :destroy]
   def index
     @items = Item.includes(:user).order("created_at DESC")
   end
@@ -24,8 +24,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == item.user_id
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
